SmartReflex: Add an adaptative method to obtain eFuse Ntargets
Since every processor has a unique eFuse ntargets value for each freq,
this method will calculate the eFuse ntarget for a given frequency basing
on the factory nvalues and pvalues. For freqs >1000GHz, it will use a
estimated slope based on the slope of the last two sections of the n and p
values functions (600-800 and 800-1000), while for lower freqs (< 300MHz)
it will use the same slope as the first section (300-600).